A lot has changed.
Moved rendering to it's own System in anticipation of rendering needs
down the line.

Created Dimension struct which holds all the chunks and has some logic
to handle persisting them to disk and generating new ones on demand.

Created volume module that mainly contains the Sphere shape. This allows
iterating over the integer points in a Sphere around some point. This
will be used in conjunction with dimension to check all the chunks in a
Sphere around the player.